requirements.txt
typer>=0.12,<0.13
click>=8.1,<8.2
rich>=13.7,<14
pypdf>=4.2,<5
pdf2image>=1.17,<2
pytesseract>=0.3,<0.4
Pillow>=10.4,<11
google-cloud-documentai>=3.7,<4
google-cloud-storage>=2.17,<3
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
qdrant-client>=1.9,<2
openai>=1.30,<2
python-dotenv>=1.0,<2
tqdm>=4.66,<5
